Capture existing multi-select list box behavior with more tests
According to the spec [1], if a choice field has both /V and /I entries
and the /Opt items they identify are different, then the /V entry
should be used. Add tests to show that the current implementation does
not behave consistently with the spec.

[1] Section 12.7.4.4 "Choice Fields" of the ISO 32000-1 standard.

Bug: pdfium:1505
Change-Id: I3564812f30d6aa693f3e5effb25f8f0fc06d6485
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/68631
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>