Enable test for formcalc eval().
Then fix the bug in formcalc translation where we were appending
an extraneous NUL onto the end of a WideTextBuf. WideTextBuf's
methods give perfectly valid strings or string views based upon
the accumulated length, so this additional NUL was being passed
into JavaScript as part of an expression.  The expression then
failed to parse when the NUL was seen, giving an invalid token
error.

- Remove several other occurrences of the same pattern.

Change-Id: I4bb68e573c9e41ecb5eca73b009ce8876238997e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/69090
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>