Adding API to check if an option in a choice form field is selected.
This change adds a public API to determine whether or not the option at
|index| in |annot|'s "Opt" dictionary is selected. The API
FPDFAnnot_IsOptionSelected() is added in fpdf_annot.h and is intended
for use with listbox and combobox widget annotations.

This CL also includes tests to validate the API for listbox and combobox
widget annotations.

Bug: pdfium:1526
Change-Id: I4bd2ae6805ef348f1c34dca7e1136824490314f4
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/69790
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Mansi Awasthi <maawas@microsoft.com>