Handle Enter/Space key events for read-only check box and radio button.
Currently if the user tabs onto a read-only check box or radio button,
pressing any character or just space/enter checks/unchecks the controls.

This CL adds a check for read-only flag in the PWL control (Check Box/
Radio Button), before processing the OnChar event.

If the window created for the control is read-only in the CFFL_Checkbox
and CFFL_RadioButton, altering the state of the control is not allowed.


Bug: pdfium:1431
Change-Id: I6fcf8de0c01d973157641fde9bc2579483878484
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/66832
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Himadri Kakar <hikakar@microsoft.com>