Reset context after running CXFA_TextLayout::LayoutInternal().
LayoutInternal() runs in a loop, but the conditions for when to reset
the context is not reachable. The caller to LayoutInternal() loops over
index values [0, N), but the check for when the index reaches N is in
LayoutInternal(). As a result, the context never resets in a timely
manner and instead resets when CXFA_TextLayout is destroyed. At which
time, CXFA_LoaderContext may have members that are dangling pointers.

Fix this by resetting the context after all the LayoutInternal() calls
finish.

Bug: chromium:997412
Change-Id: I17685e1cce951f35c8b70f8ee820ad81a5cbb209
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/70517
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>