Add a test to show FPDFImageObj_GetBitmap() does not handle masks.
Make a copy of testing/resources/pixel/matte.in and the corresponding
generated PDF for use by embedder tests. Add a test to extract the
images from it. Show that they are all the same, even though they are
different when rendered in the PDF.

Bug: pdfium:1554
Change-Id: Iced360086b21c0d93a3baeebb95615546c8e6fa3
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/71094
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>