Add CXFA_Document::GetHeap().
Since CXFA_Document creates the various CXFA_Nodes, this will be a big
help in allowing them to be GC'd down the road. The heap is reachable,
but is a couple of hops away, so add the function here rather than
making callers do these steps over and over.

-- Move some forward declarations even more forward.

Change-Id: Ic589bf7d4a4c47cc69344226ffd1c48def3585be
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/71390
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>