Remove CFWL_NoteDriver::ClearEventTargets().
It doesn't change any logic, as it only removes invalid events,
and these are not dispatched to anyways in SendEvent(). It does
return some memory, but this is only called during document
cleanup, so the memory is about to be returned anyways.

Follow-on from https://pdfium-review.googlesource.com/c/pdfium/+/72016

Change-Id: I734fbf7cf63ca47251ed0856a37e10006d72d3d5
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/72021
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>