Introduce template GCedTreeNodeMixin<>.
This mixin class is required should we attempt to make CXFA_Node
be garbage-collected, since it is a subclass of CFXA_Object which
is already garbage collected. This is the Oilpan technique for
adding a second garbage-collected class to an already garbage
collected class.

Change-Id: Ic3ecfdfb8103bac7ccfbea17b674a5ab416f951b
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/72291
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>