Remove alternate spelling of j#61v#61script from js.pdf
We suspect it makes some AVs erroneously suspect that we are up
to no good, and flag the file as suspect. It is usually best to
just avoid the possibility of the warning, since some other AV
will likely flag it again in the future. The unit tests for
PDF_NameDecode() give sufficient coverage for this case.

Change-Id: Ia186152167cafe17adcd7deeb72357bf1db41d17
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/72331
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>