Track CJX_Objects in CFXJSE_Engine::m_mapObjectToValue
CJX_Objects are 1:1 with the CXFA_Nodes that are currently
tracked in this map, but at a lower layer. These are more
appropriate for CFXJSE_Engine, especially since the data slot
on the V8 side points back to a CJX_Object, not a CXFA_Node.

Change-Id: I79d293362c8324df0b01f8e1a9087a7003f5d61c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/72570
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>