Remove CFWL_WidgetMgr::SetParent().
It is only called in one place, right after a widget is created. The
widget constructor does an InsertWidget() with the same parent and
widget arguments that will subsequently be passed to SetParent(). The
result is two calls to AppendLastChild() with the same arguments, with
the second one being essentially a no-op.

Change-Id: Ibee59582fe81e72e37e6fa76d73930edc35c004c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/72710
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>