Fix visibility on //xfa:xfa_warnings.
The config had a `visibility` set on it that had no
affect. Fortunately for it, GN wasn't enforcing
visibility of configs.

Less fortunately for it, the latest GN does, so we
need to fix this before we can roll to the latest GN.

Bug: crbug.com/1119933
Change-Id: I4ea21ed9d9a7d17a3e7a98b03818676f0722e87f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/73010
Auto-Submit: Dirk Pranke <dpranke@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>