Introduce CFX_GEModule::PlatformIface::CreateDefaultSystemFontInfo().
Replaces SystemFontInfoIface::CreateDefault(), which could not be
implemented on android and was stubbed.

The functional change are
- FPDF_GetDefaultSystemFontInfo() on android should now work.
- FPDF_GetDefaultSystemFontInfo() takes user-paths into consideration.

The lack of user-paths meant that the object returned from this
function would reference a different set of fonts that the one
we actually installed by default at creation time.

- Centralize setting default in CFX_GEModule itself.
- Update a stale comment while we're at it.

Change-Id: I621cc2e9e1ddcb12c0b843414bbefae3a0302bbe
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/73033
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>