Initialize CPPGC heap with kNoConservativeStack.
Otherwise, we might encounter a partially-constructed object,
which is conservatively scanned, rather than Traced. This works
fine so long as all the pointers are inline, but not with
std::list<> etc. where they may be out of line.

Roll back the protections introduced earlier, as tracing such
an object was in fact a v8 bug.

Change-Id: If8d2373b3c91acebe7697ff4760eed97f3efb109
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/73254
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>