Enable FPDFFlattenEmbedderTest.BUG_889099 for Skia/SkiaPaths.
When Skia/SkiaPaths is enabled, FPDFFlattenEmbedderTest.BUG_889099 fails
due to Skia's own way of hairline handling along the edges of a
rectangle. Since the discrepancy between AGG's and Skia's rendering
results is not obvious to human eyes and it goes away as the scale of
rendering goes up, we consider Skia's rendering result acceptable.

This CL adds the expected checksum for using Skia/SkiaPaths so that the
test can pass.

Bug: pdfium:1500
Change-Id: If6129849a34a0048d295eb216df3f69fb6634b2e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/73334
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Hui Yingst <nigi@chromium.org>