Remove CFGAS_FontMgr::EnumFontFromFiles.
It can't work under chrome and its sandbox process model.

It is only hit once under pdfium_test, and that occurs in
FPDFUnavailableSysFontInfoEmbedderTest where we don't want
any fonts to be found in the first place, so probing doesn't
impact success.

We've already adapted the subordinate CPDF-based font mgr to
cope with the chrome environment, so let CFGAS just piggy back
off of that.

Change-Id: I16ba3f52084e65b7ae229088b124d77c191900c5
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/73335
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>