Enable FPDFFlattenEmbedderTest.BUG_896366 for Skia.
When Skia/SkiaPaths is enabled, FPDFFlattenEmbedderTest.BUG_896366 fails
due to Skia's own way of hairline handling along the edges of a
rectangle. Since the discrepancy between AGG's and Skia's rendering
results is not obvious to human eyes and it goes away as the scale of
rendering goes up, we consider Skia's rendering result acceptable.

This CL adds the expected checksum for using Skia/SkiaPaths so that the
test can pass.

Bug: pdfium:1500
Change-Id: Idf6d7cffcf3e5709ffadca9053cac4afd3924d64
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/73491
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Hui Yingst <nigi@chromium.org>