Garbage collect CXFA_LocaleMgr.
Although not strictly required, this replaces a raw pointer to
a node with a traceable member.

In turn, pass LocaleMgrIface to the string formatter only on calls
that are affected by it. This prevents a lower layer that knows
nothing about garbage collection holding an unowned pointer to
a GC'd object.

Bug: pdfium:1563
Change-Id: Id6a372136f171a307b764a95d478c108cd3be5e4
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/73990
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>