Add pdfium::span::front() and back()
Something similar was done in
  https://chromium-review.googlesource.com/c/chromium/src/+/1539048

but we can simplify for pdfium since it only implements dynamic spans.
Add tests, renaming test suite per the suggestion in
  https://github.com/google/googletest/blob/master/googletest/docs/advanced.md#death-tests-and-threads

but do not go as far as to specify
  ::testing::FLAGS_gtest_death_test_style = "threadsafe";

as that adds 650ms per test run on my workstation.

Change-Id: I39b4402ba53471bc2a7ce0f2bc050e8d93822a04
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/74272
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>