Tidy CFX_Renderer.
CFX_Renderer is a helper class for CFX_AggDeviceDriver. It contains a
function pointer member with a confusing name. Add a typedef for the
function pointer type, and give the member a name that is consistent
with the class's member naming style.

Also move Init() to the top of the class declaration, and mark most
methods as private.

Change-Id: Iffb9e624981339bb62828df0b44b51c9ba3b0f00
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/75312
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>