Enable FPDFProgressiveRenderEmbedderTest.RenderHighlightWithColorScheme for SkiaPaths.
For FPDFProgressiveRenderEmbedderTest.RenderHighlightWithColorScheme,
due to the reasons stated in
https://pdfium-review.googlesource.com/76030, Skia's and SkiaPaths' text
rendering results may have small but acceptable discrepancies compared
to AGG's. Also due to Skia's hairline rendering method for drawing
rectangles, there are also acceptable discrepancies along the edges of
the rectangle when rendering with Skia/SkiaPaths.

This CL updates the expected results for Skia and SkiaPaths for this
test. However due to the current Skia result is not rendering the black
highlight area correctly, this test is only enabled for SkiaPaths.

Bug: pdfium: 1500
Change-Id: Ibf8371108e13d289731f91c4e630373ba748c896
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/76693
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Commit-Queue: Hui Yingst <nigi@chromium.org>