Fix SEGV in chrome under XFA for xfa_node_caption.pdf
This renders properly in the pdfium_test utility, because it
can find a suitable font amongst those we provide it with. But
chrome passes a different set of fonts, where we trip a SEGV
in CFX_RTFBreak::GetDisplayPos() if a font wasn't found.

Change-Id: Ic8cd5d77e3521fa11390b4afb6a1ab6803b44c3c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/76792
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>