Allow reproduction of the issue in 1156170 with pdfium_test.
Pump the message loop more frequently to expose GC concurency
issues. Does not fix the linked bug but rather makes it reproducible.

To avoid a number of new #ifdefs at various layers, encapsulate
the new logic in a std::function<> callback that can be passed
down to called functions.

Allows removal of a longstanding TODO() in foreground_task.in as
this test now completes as designed.

Bug: chromium:1156170
Change-Id: I82d95d71f31f9c427f330dbf08ab62ba94078079
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/76890
Reviewed-by: Hui Yingst <nigi@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>