Add FPDFAnnot_GetBorder() API
This is similar to FPDFAnnot_GetLine(), but a /Border key is valid for
all annotation types.

The array can have an optional 4th element, but that itself doesn't have
a fixed size, so it's not exposed by this API -- could be done by a
separate function.

Change-Id: I796c2ea17be1f039a6c6129dbde0b0a22276e19a
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/77010
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>