Fix `zero_area` flag in CFX_PathData::GetZeroAreaPath().
`zero_area` in CFX_PathData::GetZeroAreaPath() should indicate that when
`m_Points` is a palindromic, it forms a thin line that covers zero area.
However the assigned value for this variable represents the opposite.
This CL fixes this issue and leaves the rendering process intact.

Change-Id: Ia2ccbfb4d39691590723ed7a4f43b2e7c99892af
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/77872
Commit-Queue: Hui Yingst <nigi@chromium.org>
Reviewed-by: Daniel Hosseinian <dhoss@chromium.org>