Remove dubious combination of XFA_VT_* values
Although the XFA_VT_* values are implemented as bit-fields, they are
never accessed via bitwise operations, but as direct comparisons
usually in switch statements. Remove one place where the values
are combined, because the net result is switching against an
invalid value, which is handled in a default: clause the same way
as NULL is by both callers of the function.

Change-Id: I2f1d5e5be31115711a7c6d6fd8da09e4ae69065b
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/80716
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>