Fix subtle bug in CFGAS_GEFont::m_CharWidthMap.
It is currently a map onto int32_t. Originally, it was a map onto
uint16_t (implemented as CFX_WordDiscreteArray, which was a typedef
for CFX_DiscreteArrayTemplate<FX_WORD>, and FW_WORD was a typedef
for unsigned short). Some of the logic in GetCharWidth() relies
on storing a -1 into this map, but getting back 65535 when read,
which can't happen with the wider signed type.

Broken at https://pdfium-review.googlesource.com/c/pdfium/+/2953.
The result is logically correct, but we no longer avoid lookups
that could have been avoided.

We'll need to re-write this to something saner presently.

Change-Id: I29275f5fdb8185c602f88402962aa31c407df585
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/81450
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>