Make fixed-point arithmetic in CStretchEngine more obvious.
In particular, allow adjusting the fractional bits with one constant.
However, I expect there may be an overflow at line 612 if this is
made larger. Theoretically, we should be able to use up to 24 bits
if we move to unsigned types and watch for overflows.

Change-Id: I9a4210f15c362801c51b619e790bc152a8282b51
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/81751
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>