Return int directly from CWeightTable::GetValueFromPixelWeight().
Currently, we return a potentially NULL pointer which needs to be
checked inside a very tight loop. This should never happen if we have
set up m_SrcStart and m_SrcEnd consistently with the number of weights
that we have allocated. We still keep the bounds checks in place for
safety, and return 0 in that case. Disassembly of StretchVert() in
an x64/Linux release build gets about 100 bytes shorter.

Change-Id: Ic187c3e514071550bd2c7bc0b4cbd43df2c4d13c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/81930
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>