Validate more inputs in CPDF_Function::Call().
Per table 38 in the PDF spec, for each pair of values in the domain, the
first value has to be less than or equal to the second value. Validate
this and reject functions with invalidate domains. Do the same for the
range.

This avoids passing values to pdfium::clamp() where the "low" value is
higher than the "high" value. Since pdfium::clamp() is now compliant
with std::clamp(), the extra validation from this CL avoids undefined
behavior.

Bug: chromium:1235105
Change-Id: I548a36c009bb4267aadf127db2346aa85f1e9384
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/83810
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>