Remove enum BarcodeType.
While this enum provides an abstraction between the types of
barcodes that XFA might know about vs. the types of barcodes
actually implemented by the current barcode library, it is unlikely
that we would ever replace the later, so the use of its values
throughout XFA isn't much of a layering issue.

Change-Id: I47c7bca783b41c820e969754d734e76c24df95bb
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/84356
Reviewed-by: Hui Yingst <nigi@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>