Remove non-const form of CPDF_ColorSpace::AsPatternCS().
The only caller only requires a const version, so we can save one
slot in N subclass vtables. Also remove a NOTREACHED(), because
although it may be true in practice, in theory one should be able
to call this on any CS and check for null return.

Change-Id: Idee097f732590474d35a1033228faf6b3dce13cd
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/84713
Reviewed-by: Hui Yingst <nigi@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>