Make GetEstimatedImageSize() a method of CFX_DIBBase
Rename to GetEstimatedImageMemoryBurden() since the result is
used in cache limit calculations. Allow the base class to return
the size of the palette for better calculations when there is no
buffer. This is a second step closer to moving to a world where
GetBuffer() is only called on CFX_DIBitmap, not the base class.

Change-Id: Id01c148fbfe204753eba0ecd0566df101e606043
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/85691
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>