Fix span calculation in CPDF_DIB::GetScanline()
Ideally, the span size would match the calculated src_pitch_value, but
some spans are actually larger than we think. In cases when there is a
decoder or a cached bitmap, rely on the bounds being correctly set from
those spans, rather than recalculating them based on the current notion
of pitch.

Try to return spans representing only the initialized portion of a
buffer when we copy into one.

Having landed this CL, we should be able to re-enable the spancpy()
in CFX_DIBBase::Clone() that was reverted in a prior CL.

-- take a branch out of a for-loop while at it since I was confused
   for a moment if both halves advanced the same amount.

Change-Id: I557880b610af75f8bc68592d462b05cd8be964f2
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/85851
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>