Update test case for bug 665467.
Rewrite the PDF so that it is syntactically correct and can render in
PDF viewers that are not PDFium-based. Since this required switching
fonts to better match the original PDF on the bug report, update the
test expectations to match.

Partially based on https://pdfium-review.googlesource.com/87150

Bug: chromium:665467
Change-Id: I27caaa08c408d8e5e86a7be0be3508a8c46c2646
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/87210
Reviewed-by: Hui Yingst <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>