[Skia] Fix the issue that wrong scale is applied to y position.
If a glyph's transformation matrix is [a b c d 0 0], then `a` is
the scale factor applied to its x position, `d` is the scale
factor applied to its y position, and `b`,`c` are used for skewing
the original glyph bitmap.

However, when rendering bug_1402, cp.m_AdjustMatrix[1] (AKA `b`) is
used for scaling the glyph on the y axis, which happens to be 0 and
erased the glyph's y position. This resulted in all characters with
the same transformation matrix are rendered lower than expected.

This CL fixes the issue by apply `-cp.m_AdjustMatrix[3]` (`d`) as
the scale factor of a glyph's y position, which takes consideration
that the y position is inverted from character to device.

Bug: pdfium:1740
Change-Id: I5ed3d2f36562564ebf1039ec18a7354adee2af87
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/87410
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nigi <nigi@chromium.org>