Rename fm2js dir to formcalc.
For some reason, the FormCalc to JS code is in a directory called
"fm2js". Rename to it formcalc to simplify the name. The files and
classes in the directory still have "fm" in their names, and the
associated fuzzer targets retain their old names temporarily to give
Chromium a chance to switch over.

Also rename some variables that have "fm2js" in their names.

Bug: pdfium:458
Change-Id: Ie6375ce0f79638cca6744fdeeaac47928a857c63
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/89852
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>