Fix PathService::GetThirdPartyFilePath() to work in a Chromium build.
When running pdfium_embeddertests as part of a Chromium build, it tries
to use Chromium's //third_party directory instead of PDFium's. Fix this
by checking for the presence of the bigint sub-directory, which is
in PDFium's third_party dir and unlikely to be in its embedder's.

Change-Id: Ie3633c562eeef77e62d26c111e53bbc3a36dad75
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/91271
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>