Remove unreachable CBC_QRCoderMode::sKANJI code.
In CBC_QRCoderEncoder::Encode(), the encoding is always "utf8", so the
check against "SHIFT_JIS" is always false and "sKANJI" is never used.
Delete the associated code, and nearby functions that take the encoding
as a parameter but never use it.

Switch to a range-based for-loop along the way.

Change-Id: If5bafad9b9aada3f29ec2df0831d34d679d3131f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/91710
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>