[Skia] Add a test case for blending mode used for image masks.
There is no existing pixel test that provides test coverage on the
`SkBlendMode` used in SetBitmapPaint() when image masks are not
completely opaque.

This CL adds a new test based on pixel test bug_591137.in in which the
characters are transparent and overlap with each other to prevent future
regression in this function.

Bug: pdfium:1746
Change-Id: I4165040313274644d7302a2e0b85b1046a643107
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/91750
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nigi <nigi@chromium.org>