Remove IPDFSDK_AnnotHandler::ReleaseAnnot().
For the CPDFSDK_Annot subclasses that need to do additional work before
being deleted, do that work in their destructors.

Also change CPDFSDK_PageView::DeleteAnnot() to do all the work, instead
of letting its caller do some of the work. Rename it to
DeleteAnnotForWidget() to better reflect what it does now, and do not
return a bool, as the caller does not check the value.

Change-Id: I09eebd597d0d47adbba7945e919630ce747718c1
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/92131
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>