Avoid infinite loop caused by floating-point loop counter.
When a floating-point loop counter is incremented/decreased by an amount
that is too small to change its value given its precision, the loop will
get stuck and cause the process to hang.

This CL changes vcgen_dash::calc_dash_start() to decrease `ds` by a
multiple of the total dash length first, so that `ds` won't be too big
to cause an infinite loop.

Bug: pdfium:1822
Change-Id: I3750f8f59260df2c06fd734edce4d57535db8e12
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/93071
Commit-Queue: Nigi <nigi@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>