Run non-JS /OpenAction even if JS not present.
Very curious since it has been this way since before the
beginning. Might have been cut/paste from previous method
which invoked a JS action, but seems like OpenAction should
not be predicated on that.

Unfortunately, the test harness isn't equipped to deal with
the consequence of this change. The follow-on CL, however,
requires it and will enforce behaviour.

Change-Id: I475e4b6bb11045918cd2b6b3b0185ef6fc25e40b
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/93370
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>