[Skia] Avoid rendering adjacent characters connected to each other.
If a glyph's default width is the same as its width defined in the PDF,
we should avoid using the device driver to render such a glyph so that
the adjacent glyphs won't be connected to each other. Instead, render
these glyphs with their adjustable paths.

Bug: chromium:736703
Change-Id: I115773beaa64031d982f362f946d19ec68c126fc
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/93890
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nigi <nigi@chromium.org>