Make checking out Skia and V8 optional.
In the root DEPS file, add variables for controlling whether Skia and V8
get checked out or not. Then set the default values for these variables
based on where the variable checkout_configuration is set to 'minimal'
or not.

This allows PDFium embedders that just want the most basic build
configuration to avoid ~1 GB of unnecessary downloads.

Bug: pdfium:1852
Change-Id: I54a517e5e9ae196cdbd48347738d61ca8c748ec8
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/95571
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>