Fix path to pdfium_partition_alloc target.
The path should be relative, so it works correctly when PDFium is
embedded in another project. As is, the absolute path will look for
pdfium_partition_alloc in the top-level third_party directory. When
PDFium is embedded, that third_party directory belongs to the embedder.

Change-Id: I991e02b570eb3c2f88a7fb8d79c248c921de2de0
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/96770
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>