Introduce Skia is renderer queries
Add query functions to determine if a renderer variant that uses Skia
is in use.  This is in preparation for enabling Skia and AGG renderers
to coexist in a single build and for the renderer to be selectable at
runtime.

These query functions enable tests to be incrementally refactored to
always include Skia and AGG definitions in a test build, and to select
at runtime which results to use for expectations.

For now these new queries are predictably a direct match of the Skia
defines.  Enabling this to be runtime-driven based on a variable will
come as more changes to the underlying renderer support are filled in.

Bug: pdfium:1878
Change-Id: Ie3e74b3ac4355ad90cb478d707356331af52bee7
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/96733
Commit-Queue: Alan Screen <awscreen@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Nigi <nigi@chromium.org>