Return retained objects from the indirect object holder, part 1.
We've long considered these objects as permanent for the lifetime
of the document, but there may be ways to trigger a parse, for
example, replacing the object - either in error, or with one of a
higher generation number. So extend the lifetime of the objects
we return to callers.

Change-Id: Ia343f040b48a316d9d0c4513561f2bc0cd45fd97
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/97310
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>