[Skia] Set the correct group knockout status for fill-and-stroke paths
When painting a fill-and-stroke path, the group knockout status needs
to be set to true to avoid the unwanted border overlapping rendering
effect. This CL sets group knockout status to true for Skia before
painting such paths and then restores the original status after the painting is done.

Bug: pdfium:1802, pdfium:1651
Change-Id: I9c7cc32f4af3fe64fe0b8f10305b3ef64ed5cdc9
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/97350
Commit-Queue: Nigi <nigi@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>