Return const pointer from CPDF_Stream::GetInMemoryRawData().
Callers to CPDF_Stream::GetInMemoryRawData() do not need the non-const
pointer to the data. GetInMemoryRawData() was only returning a non-const
pointer because fxcrt::MaybeOwned required it for the owned-data case.

In CPDF_StreamAcc, replace its MaybeOwned usage with absl::variant. Then
the variant can store either a read-only span, or some kind of owned
memory. Then CPDF_StreamAcc no longer need to have separate data and
size members.

As part of this CL, also restrict when GetInMemoryRawData() can be
called, and update CPDF_StreamAcc to match that behavior. Update the
comments for GetInMemoryRawData() to better explain how it is intended
to be used.

Change-Id: I6a05e8b5314434447e45c44596c81771765756fb
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/97535
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>