Add a libcxx_revision presubmit check.
Both DEPS and buildtools/deps_revisions.gni have libcxx_revision, and
they have to stay in sync. Add a presubmit to make sure this is the
case.

The check needs to be a separate tool that lives outside of
PRESUBMIT.py, because the ReadFile() API available inside of
PRESUBMIT.py cannot read files outside of the PDFium git repo. Since
buildtools is pulled in via DEPS, buildtools/deps_revisions.gni is
actually in a different git repo.

Change-Id: I5eb2aff800bf89da2d17045be61bbbf61e9cab4e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/97773
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>