Explicitly allow IFX_WriteStream::WriteBlock() calls with size of 0.
Make it clear that the interface allows this. Fix implementations to be
consistent with the interface, and remove a few checks that skipped
WriteBlock() calls because the size is 0.

At a test case to make sure CFX_MemoryStream::WriteBlock() works as
expected.

Also consistently not return false in WriteBlock() implementation when
the data pointer is null.

Change-Id: I4962f56b66c31c53339feb120f5a33d440fa72d7
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/98090
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>