Remove dead code in CPDF_TextPage::Init().
The loop inside CPDF_TextPage::Init() checks to see if `m_CharIndices`
is empty. It is never empty because the only way to enter the loop is to
first add an item to the vector. The loop also never removes any
elements from the vector.

Change-Id: I4274a50de1fc3832727aa650606cdf808fe6e9df
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/98290
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>