Create helper function and move a constant to a header file
Move the process of creating a palette buffer from a 1-bit-per-pixel
bitmap into helper CFX_ImageStretcher::BuildPaletteFrom1BppSource()
and move constant `kPaletteSize` into cfx_dibbase.h so that they can
be reused for Skia rendering process in the future.

Also in this helper function, adds DCHECKs for the two palette colors' alpha values (They are set to 255 during the palette construction).
This helps avoid the calculation of `a` within the for-loop.

Bug: pdfium:1810
Change-Id: Id5a4b94097002c5811a5cbb23ff450da25b50869
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/98330
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nigi <nigi@chromium.org>