Fix interaction between RetainPtr<T> and transparent comparisons.
The key is providing an implicit operator T* so that we can make
comparisons in terms of raw pointers in all cases. Unfortunately,
this now allows passing RetainPtr<T> as T* without an explicit
Get(), which is cleaner but will not catch the use of unretained
references.

-- remove now-redundant comparison functions.
-- kludge around some ambiguity in CPDF_DictionaryLocker

Change-Id: Ib06f665361f4efa514c6f028eb24af114f7506dc
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/98394
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>