[Skia] Create a new palette when upsampling 1-bit-per-pixel bitmaps
For a 1-bit-per-pixel bitmap, its color palette sometimes stands for
the boundary of a range of colors instead of just 2 colors.

In Upsample(), this CL adds the process to create a new palette based
on this range and uses it for upsampling the bitmap to 32 bits per
pixel.

Meanwhile, move the process that fills a 32-bit-per-pixel buffer with
the palette into a helper Fill32BppDestStorageWithPalette() so that
this process can be reused within Upsample().

Bug: pdfium:1810
Change-Id: Ia2c4394d8d7d0809a3987d5a3304e7c7892ade6f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/98175
Commit-Queue: Nigi <nigi@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>