Rework CPDFSDK_FormFillEnvironment::JS_appResponse() callers.
Since JS_appResponse() returns how much of the buffer is actually used,
and the function that handles the returned buffer also takes a size,
there is no need to pad the buffer with NULL characters. Remove the code
that assumes this is the case.

Change-Id: Ia83105679a82bf14c388c6e638438b5d86b3503a
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/98571
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>