Enable FPDFEditEmbedderTest.SaveAndRender for Skia
CL[1] fixed the issue that paths with transparent fill and stroke
colors are rendered with overlapping area of the two colors and the
Skia rendering result for FPDFEditEmbedderTest.SaveAndRender is now
acceptable.

This CL adds the Skia expectation for this test so that it can be
enabled.

[1] https://pdfium-review.googlesource.com/97350

Bug: pdfium:1651
Change-Id: Ide89e8f1c6c5d22bdac90f6649b15bb5df2a68f1
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/98631
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nigi <nigi@chromium.org>