Make CPDF_FormField::GetFieldAttr() a non-static method
Most callers want to invoke this on the field itself, so we can avoid
the step of passing a dictionary back to the caller. Rename existing
method to GetFieldAttrForDict() to handle odd cases where there isn't
a form field to directly supply the dict.

Change-Id: Icfcde3cfc0c02a595ae9f9119ba24eb24dec6a71
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/98790
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>