Fix overlapping annot rendering with focused annot.
The reason CPDFSDK_AnnotReverseIteration existed was to put the focused
annotation at the end of the list, so it will draw on top of other
annotations. When https://pdfium-review.googlesource.com/98870 fixed the
drawing order for overlapping annotations, it broke this use case.

Add CPDFSDK_AnnotIteration::CreateForDrawing() as a factory function to
create CPDFSDK_AnnotIteration objects that iterate correctly for the
drawing use case. Use it in CPDFSDK_PageView::PageView_OnDraw() to fix
annotation drawing. Make sure it works as expected with a new test case.

Bug: chromium:1372651
Change-Id: I0bf673e9af9eb59773f7ef933fc363aab9f22abc
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/99110
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>