Fix CHECK failures with GetFontData() calls.
In an ideal world, like CL [1] assumed, calling GetFontData() twice in a
row would yield the same return value both times. In the real world,
this is not the case and sometimes the call to get font data fails.
Handle the errors that occur in the real world in various CFX_FontMapper
methods that call GetFontData().

Bug: chromium:1372234

[1] https://pdfium-review.googlesource.com/97976

Change-Id: I5bfa6e234c6a2b23a9ef8cf48c31a1a38a7e83ef
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/99351
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>