Actually fix clang-tidy report
Returning a retained reference is always safer, although this
is not reachable in practice. The clang-tidy report fires when
a function returning a retained reference is assigned to a bare
pointer because it can't deduce that there are other references.
When the temporary holding the retained result prior to conversion
goes out of scope at the next semicolon, the pointer outlives it.

Fixed: chromium:1380478
Change-Id: Iaa582b40c2f6a8d4fb227a2437e4a1635c618521
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/100612
Reviewed-by: Lei Zhang <thestig@chromium.org>
Auto-Submit: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>