Populate bitmaps even when using SkPictureRecorder
Populates the "m_pBitmap" and "m_pBackdropBitmap" members of
CFX_SkiaDeviceDriver, even when recording to an SkPictureRecorder. Some
code paths otherwise are skipped, reducing the fidelity of the SkPicture
output mode.

Since SkPictureRecorder isn't a rasterizing backend, these bitmaps are
filled with a solid color, in an attempt to make it more obvious when
their (meaningless) contents are used.

Bug: pdfium:1929
Change-Id: I7a73bc23eec71a4f3c0fd81ee6968ee0fe3c4aab
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/101452
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: K. Moon <kmoon@chromium.org>