[Skia] Suppress failed corpus tests
When rendered with the `--render-oneshot` option, the results of
form_combobox_plus.pdf and form_combobox_product.pdf are slightly
different compared to when being rendered without this option. This is
causing corpus tests to fail on Skia bots. Suppress these tests until
this issue is fixed.

Bug: pdfium:1855, pdfium:1963
Change-Id: I3ddc22509c45db018a9270c8d53772a14ebc16ac
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/95950
Commit-Queue: Nigi <nigi@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>