Replace CFX_ScanlineCompositor::m_iTransparency with 2 structs.
The `m_iTransparency` integer is overloaded and has different meanings
in different parts of CFX_ScanlineCompositor. Replace it with an
absl::variant to better distinguish what it actually means. Then for the
2 different contexts in which it is used, replaced the int value with a
struct, so it is easier to understand what the value stands for.

Change-Id: I27dae274e0487124e30ff537d4cb9976ba1067f4
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/102672
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>