Add timeouts to test_runner.py
Adds a (5 minute) timeout to subprocess calls in test_runner.py, to
ensure the test runner eventually terminates even if a subprocess hangs.

Notably, this does not add a timeout to RunCommandExtractHashedFiles()
in common.py; this will be addressed in a future change.

Also changes the test log to a bytes object, instead of a str object,
allowing the command output to deviate from valid UTF-8. The
TimeoutExpired exception does not decode captured output, and ResultDB
does not require the test log to be valid UTF-8.

Bug: pdfium:1964
Change-Id: I8df8d6df9eee7f475cef8cadb6246c782cbf2bb3
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/102890
Commit-Queue: K. Moon <kmoon@chromium.org>
Reviewed-by: Nigi <nigi@chromium.org>