Fix an integer overflow in CFX_Path::GetBoundingBoxForStrokePath()
CL [1] allows a move operation that closes the path to be rendered,
which means points with "move + open" combo are no longer the only
moves to be anticipated in CFX_Path::GetBoundingBoxForStrokePath().
Due to [1], integer overflow is triggered when a "move + close" point
is the first point of path.

This CL makes sure "move + close" points are handled when updating
boundary of stroked path.

[1] https://pdfium.googlesource.com/pdfium/+/2ef13c3534c101c5267e48442b60b21ca8409d03

Bug: chromium:1414373, chromium:1414488
Change-Id: Ie11d1269bf116ac3f6c0290e7287c6e3c1ca4a58
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/103691
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nigi <nigi@chromium.org>