[Skia] Remove redundant DrawDeviceText() branches
Removes redundant branches in CFX_SkiaDeviceDriver::DrawDeviceText(). It
turns out that most of the logic in DrawDeviceText() is unreachable,
because TryDrawText() already handles those cases.

We can draw runs of text in Skia in one of three ways:

1. Translating (positioning) glyphs.
2. Rotating, scaling, and translating (RSX form) glyphs.
3. Painting each glyph separately with a full transformation matrix.

These get progressively less efficient, so we try to use the faster
methods first. It turns out that TryDrawText() covers the first two
cases, so DrawDeviceText() should not be handling them, too.

Note that the condition in DrawDeviceText() for one-at-a-time mode was
strictly a superset of the condition in TryDrawText(), so any case
treated as one-at-a-time by TryDrawText() also was treated the same way
by DrawDeviceText().

Fixed: pdfium:1963
Change-Id: Ie9e1fdc7567aa7d983fd41d16fd9c067a8d7c622
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/104651
Commit-Queue: K. Moon <kmoon@chromium.org>
Reviewed-by: Nigi <nigi@chromium.org>