[AGG] Add more pixel tests for out-of-bounds coordinate lines
Add minimal test cases for line misrenderings when the coordinates are
out-of-bounds. The bounds are [-32000, 32000]. The expected result is
the Skia rendering, while the buggy AGG rendering result is suppressed.

bug_1846.in has been renamed to bug_1972_1.in so that the similar line
tests are grouped together.

Bug: pdfium:1972
Change-Id: Ia83198cd8d478ac39ad2ce0b56dc79371b1a13b4
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/104515
Commit-Queue: Andy Phan <andyphan@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>