Add a new test for rendering JPX images
Adds a new test case that the *Alpha images contain soft-mask
information, but the PDF should not render that mask due to missing
/SMaskInData entries. This will help improve test coverage on images
of which the image component numbers do not match the color channel numbers.

The expected result reflects the correct rendering result, which
aligns with the Acrobat rendering result.

Bug: pdfium:1747
Change-Id: I5a5e70a2aab5b7ae867ac7401ce98ba5f295100a
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/104933
Commit-Queue: Nigi <nigi@chromium.org>
Reviewed-by: K. Moon <kmoon@chromium.org>