Trim third_party/base/bits.h
PartitionAlloc is now pulled in via DEPS, and it is a standalone library
with its own copy of bits.h. This other copy of bits.h, which used to
support PartitionAlloc, is mostly no longer needed. Trim it down to the
only needed function.

Change-Id: Idd1fe917d6623d2d2a4b86aa78b7052b79e6d3c8
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/105050
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>