Add a pixel test for axial shading
Add a pixel test for axial shading. This is a simplified PDF from the
provided PDF in pdfium:2001 and it can be rendered in various viewing
tools (including Acrobat reader) without having error messages.

The expectation is exported from Preview and it should be fairly close
to what the PDFium expectation will be after this issue is fixed.

Note: No .in file is uploaded for this issue because getting rid of
the binary streams in this PDF will trigger Acrobat to report errors.
You can find a copy of the .in file at [1].

[1] https://bugs.chromium.org/p/pdfium/issues/detail?id=2001#c6

Bug: pdfium:2001
Change-Id: I497845b220fedef81987689b8ab83c78f18a1772
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/105410
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nigi <nigi@chromium.org>