Add object tree traversal utility functions
Add GetObjectsWithMultipleReferences() to figure out which objects have
multiple references. Then code that manipulates PDFs can use this
function to determine if it is safe to edit an existing object. If it is
unsafe, the code can clone before editing instead.

Also add GetObjectsWithReference() to figure out which objects are
referenced. This can be used to determine which objects have no
references and can therefore be deleted.

Bug: chromium:1428724,pdfium:1409,pdfium:2012
Change-Id: I28817068d50c79f36d12bd50c1910937241194a5
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/105611
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>