Avoid -Wcovered-switch-default errors in various places
When all the cases in a switch statement have been covered, there is no
need for the default case. Remove them as they are dead code. This
covers the majority of the remaining cases and a separate CL will try to
enable the new warning.

Bug: chromium:1393075
Change-Id: Iafd7633d62c5d405d257ad3582205a9c23b2e7cc
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/106650
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>