Fix GN visibility checks when building with chromium.
Standalone pdfium doesn't have a //base/BUILD.gn upon which to depend,
so it must depend on the partition_allocator sub-directory targets.
These violate GN constraints, however, when PDFium is built in a
non-standalone manner and //base is available.

This is currently blocking a roll of PDFium into chromium.

Bug: pdfium:2038
Change-Id: Ib2c6a84876036902d6873dedfc18e030c5aa7c6c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/107711
Reviewed-by: K. Moon <kmoon@chromium.org>
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>