Remove last reference to SkBitmap
Removes the last remaining reference to SkBitmap in the PDFium code,
which was in CFX_SkiaDeviceDriver::GetDIBits(). This has no functional
effect, but the Skia team discourages new uses of the SkBitmap API.

Also fixes a bug in which GetDIBits() was using the destination buffer's
dimensions to draw the source image. This didn't matter as long as the
source and destination dimensions were the same.

Bug: pdfium:2007
Change-Id: If2dbf45a6668d02db25a387ac6ac391c1ffc2882
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/108730
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Auto-Submit: K. Moon <kmoon@chromium.org>
Commit-Queue: K. Moon <kmoon@chromium.org>