Remove force_alpha parameter
Removes the force_alpha parameter from RealizeSkImage(), as this only
has an effect when the CFX_DIBBase is not in a mask format (k1bppMask or
k8bppMask). There is only one such case, which can be converted
trivially to use k8bppMask.

Also removes CFX_DefaultRenderDevice's SetBitsWithMask() overload, as
the implementation is equivalent to the one in CFX_RenderDevice.

Bug: pdfium:2034
Change-Id: I9f4ec7556b0132368063764b182e8a40e6e325e1
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/108710
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: K. Moon <kmoon@chromium.org>