Optimize empty map cases in CPDF_CrossRefTable::UpdateInfo()
In the case where UpdateInfo() receives an empty map, that is
effectively a no-op. Just return early and avoid moving anything around.

In the cases where CPDF_CrossRefTable has an existing map that is empty,
simply take the new map in its entirety, instead of processing it one
element at a time.

Also remove the rvalue references. They are not needed in this case, per
the style guide.

Change-Id: I05ee6f768a8a22456206674ee48d57cace5fcb28
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/109730
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Nigi <nigi@chromium.org>