Avoid misaligned loads in LoadGlobalPersistentVariablesFromBuffer().
Chrome doesn't use any of this code, but the pdfium_unittests will
trip a warning about misaligned access under a full-up ubsan build.
Fix this just for the sake of quieting the bots.

The problem is that once a fixed header is extracted from a buffer,
the remaining data is packed and can't be expected to align on machine
words. The buffer was essentially constructed by copying in this manner
rather than by assignment, so reconstruct its elements in the same way.

Bounds checking has always been dicey as it is assumed that the
data is coming from local storage as written by us.

Bug: pdfium:2066
Change-Id: Idc9edcef50cfee5899a1fb82c39a1299399382cc
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/110030
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>