Fix some DCHECK_OP signness mismatches
In Chromium, DCHECK operations like DCHECK_EQ(lhs, rhs) expects the two
operands to have the same signness. PDFium's DCHECK operations do not
have this requirement, as they are simpler implementations.
Nevertheless, make the signness of the operations match, to make it
easier to switch to a better DCHECK operation in the future if desired.

Change-Id: I8f49a2a4e7c9634df846486c62cdd1e28d9e6ac9
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/110731
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>