Allow StructElement_Attr_GetName to return buflen with null buffer
The documentation of FPDF_StructElement_Attr_GetName currently
indicates:

```
buffer             - A buffer for output. May be NULL. This is only
                    modified if |buflen| is longer than the length
                    of the key. Optional, pass null to just
                    retrieve the size of the buffer needed.
```

However, the current implementation doesn't allow the caller to pass
buffer=NULL to obtain the required buffer length.

Removing the check for `!buffer` is enough to simply populate the
buffer size and return true.

Change-Id: I8249c2fed1c07a133c66e7d38dc9ff56f7de3970
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/110711
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>