Split TextObject rect for each kPiece
When an /ActualText is used on an object, the object is represented as
text by a sequence of CharInfo marked as kPiece. Prior to this change
each CharInfo was assigned the full rectangle of the original object.
This change splits the the original object's rectangle between the
kPiece CharInfos. This allows for each CharInfo to be independently
selectable instead of all of them overlapping.

Bug: pdfium:2069
Change-Id: Iea08d4b28fa9489f159b6e0c0cc0c2d2554db19b
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/111150
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Ben Wagner <bungeman@google.com>