Avoid another PartitionAlloc allocator shim incompatibility
As reported on [1], builds on macOS do not work out of the box when
use_custom_libcxx is set to false. This build config used to work prior
to the existance of the PartitionAlloc allocator shim. Adjust the
default _use_shim value to take use_custom_libcxx into account.

[1] https://groups.google.com/g/pdfium/c/HjTzkCwOdTk

Change-Id: I1339e1fbb46db204bf79f160790298a5038b52b7
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/111814
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>