Combine CPDF_DeviceBuffer::Initialize() and GetBuffer()
Change CPDF_DeviceBuffer::Initialize() to return the initialized bitmap
on success, and declare this as part of the API contract. Then the
caller can just check if the returned bitmap pointer is null or not, and
does not have to second guess if the bitmap has a buffer.

Bug: pdfium:2047
Change-Id: I15300c7be1a0447ebcd9983249f8279f8dae1253
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/111992
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>