Always use base::raw_ptr<> implementation when PartitionAlloc is present
Now that RawPtrNoOpImpl initializes by default (modulo build args),
we can use it whenever Partition Alloc is present. This is a first
step towards minimizing some code duplication between PDFium and PA.

Also activate a simple CHECK() on BRP dangling detection. We'd not
done this in the past as some of the elegant reporting mechanism
lives chrome-side. But we need this to make some unittests work
against the new #defined symbols. This is a test-only change.

Bug: pdfium:2038
Change-Id: I5b624f82343d75cc62e362e464f88a1c759468fa
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/112350
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>