Fix incorrect font dictionary usage in GeneratePopupAP()
When GeneratePopupAP() calls CPDF_DocPageData::GetFont(), GetFont()
expects a font dictionary. GeneratePopupAP() is incorrectly passing in
the resource dictionary for all fonts. This ultimately causes the font
mapping code to ask the PDFium embedder for a font with no font name.
Fix this by generating the font dictionary and the font resources
dictionary separately. Thus making it easy to call GetFont() with the
correct dictionary.

Update corpus tests to match the new behavior. In FPDFAnnotEmbedderTest,
the generated content stream is slightly different because the font
metrics are different. This then compresses differently, which is why
the expected PDF size has changed. This change is OK, because the test
case is checking for size differences.

Change-Id: I313411864f993e096cbd41f440791bdf821bf268
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/113113
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>