Avoid setting the private tif_fd field in struct tiff
CTiffContext::InitDecoder() calls a helper function which wraps
TIFFClientOpen() and sets the tif_fd field in the returned struct. This
is a private field that the libtiff caller should not be reaching into.
The field also appears to be of no use to CTiffContext. Stop setting
this field and simplify the code. Then the code can just include the
public tiffio.h instead of the private tiffiop.h.

After removing the tiffiop.h include, adjust the bundled libtiff to
set error handlers itself.

Clean up C-style casts along the way.

Bug: chromium:1501495
Change-Id: Ic42e7c7e1af68483814f72333fd8a32547fb900f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/113594
Reviewed-by: Nigi <nigi@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>