Give "V4" and "V5" parser code better names
In CPDF_Parser and related code, there are references to "CrossRefV4"
and "CrossRefV5". But what happened to V3? It turns out V4 means PDF 1.4
and before, where only cross reference tables existed. V5 means PDF 1.5
and later, as the PDF 1.5 spec introduced cross reference stream. With
this in mind, rename "CrossRefV4" to "CrossRefTable" and rename
"CrossRefV5" to "CrossRefStream". For "TrailerV4", simplify that to
"Trailer", as a cross reference stream has a dictionary built-in and
does not need a separate trailer dictionary. i.e. the trailer is only
ever associated with a cross reference table.

Similarly rename variables with "v4" or "v5" in their names. Rename a
few other variables along the way.

Change-Id: I11eb14099cb86401fd42375c31afe5ecf7171e99
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/116291
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>