Remove CFX_Font::IsEmbedded()
CTextOnlyPrinterDriver::DrawDeviceText() is the only caller to this
method, and it does not actually care about the font. The reason it
checks whether the font is embedded is because the code started off as a
copy of another driver that no longer exists.

Note that CPDF_Font::IsEmbedded() still exists, and that is what most
code calls.

Change-Id: I47df30871c0cc4ef4a21a080e98b65e6440e770e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/116871
Reviewed-by: Thomas Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>