Convert EXPECT_STREQ(x, y.c_str()) to EXPECT_EQ(x, y).
The objective is to stop calling .c_str() except when absolutely
necessary, as extracting a c_str() from one of our string types
may someday require UNSAFE_BUFFERS(). It also may leave the tests
vulnerable to, say, injected NULs.

The gtest macros can handle these EXPECT_EQ comparisons so long
as operator==() is defined for the string type. The risk is that
a bug in operator==() might propagate to other tests, but we
test this itself in other tests.

-- Fix several places that were including a trailing NUL.

Change-Id: I9bdbd857c45b2356719a19560a30c35a80b2bff5
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/116912
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>