Change CLZWDecoder to use DataVector
Switch to DataVector instead of storing data in a unique_ptr. Adjust the
caller to temporarily copy the result to a unique_ptr for now to make
this work. Use more size_t in CLZWDecoder as appropriate, and change
ExpandDestBuf() to return a bool to indicate success.

Bug: pdfium:1872
Change-Id: I2e46945aa03887dbb04952537b9cdf76533b0284
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/119232
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>