Fix crashes in FPDFAttachment_GetStringValue()
When the key argument is "CheckSum", FPDFAttachment_GetStringValue() has
some special code to handle that case. This code path assumed the value
of the checksum will always be a string, per spec. In the wild, the
value could have some other type. Add some tests that would confuse the
existing implementation and cause a nullptr crash. Then fix the
implementation to not crash.

Change-Id: I4cc7d6a568634512fa16ffa1bea15d9713cfa00f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/119410
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>