Do not use strict mode when decoding JPEG2000 images
Acrobat Reader can render the PDF attached to the bug report, while
PDFium cannot. This is because PDFium uses libopenjpeg, and in the past
libopenjpeg only had strict mode, which rejects the image in the PDF and
refuses to do partial rendering. Now opj_decoder_set_strict_mode()
exists in libopenjpeg. Use it to unset strict mode, so PDFium can render
the PDF in the bug report in the same way as Acrobat Reader.

Bug: 42270564
Change-Id: I77b1f73659d48252d488a4a1bd170cce20017aff
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/119592
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>