Restore JPEG2000 colorspace mismatch handling workaround
The code added in CL [1] to workaround a rendering issue with
iOS-generated PDFs got removed in CL [2]. Put it back to fix what looks
like a recurrence of the bug. Unfortunately, the bug_1012369.in pixel
test case was not sufficient to catch this particular case.

[1] https://pdfium-review.googlesource.com/65753
[2] https://pdfium-review.googlesource.com/104031

Bug: 345431077
Change-Id: Ie7097830691e80877fc1632b5486c0d0033f9994
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/120213
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>