Make CPDF_PageLabel::GetLabel() faster
Implement CPDF_NumberTree::GetLowerBound(), so that
CPDF_PageLabel::GetLabel() can find the value it needs in 1 search of
the number tree, instead of N. With this speed up, enable
PageLabelTest.GetLabelPerf, which now runs in ~1 second instead of ~200.

Since CPDF_PageLabel is the implementation for FPDF_GetPageLabel(), the
public API is faster too.

Bug: 347268873
Change-Id: Ie64d94be78227b975db897411c386de1babf7065
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/120533
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>