Retire fxcrt::spanset() and fxcrt::spanclr().
Convert to fxcrt::Fill(), which is more in-line with what we expect
to be the future C++ vocabulary used in chromium (i.e. ranges::fill).

This has the nice property that we can fill various containers
without first having to make a span from them.

Change-Id: I23018969b8a47f2207adb8e40b6feb36b19ecd72
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/120772
Reviewed-by: Thomas Sepez <tsepez@google.com>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>