Simplify RenderDeviceDriverIface::GetClipBox()
Return the rect direct instead of via an out-parameter. Avoid wrapping
the rect in std::optional because most implementations do not fail. For
the ones that can fail, return the entire size of the underlying device.
This is consistent with how the callers would handle the failure case.
Now the callers do not have to.

Also mark GetClipBox() as const.

Change-Id: Ia36fa94024c1529d00604e9fcb712802655892e3
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/121240
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tom Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>