Remove UNSAFE_TODO() from CFX_DIBitmap::Clear()
Switch to spans. Then it is not so hard to fill every scanline if
needed. Instead of filling the first scanline and copying it to the
other scanlines.

To make the changes build, change fxcrt::Fill() to take a rvalue.
Otherwise fxcrt::Fill(GetWritableSpan(), val) does not compile.

To make it slightly easier to manage color component variables, add
ArgbToBGRStruct() and use that in place of ArgbDecode().

Bug: 42271176
Change-Id: Idf7a682bbc82d045922a23b2c2feb3090a58a630
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/121450
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@google.com>