Add a pixel test for /TR
Take some existing test PDFs with different image types and combine them
together into a single page in a PDF. Then create a second page in the
PDF that is identical to the first page, except everything has a
transfer function applied.

This demonstrates an issue that shows transfer functions are not
applied correctly for Mac + Skia. Accept the incorrect image as-is so
the test can pass, as the exact test expectation cannot be
pre-determined.

Bug: 351796783
Change-Id: I99c9da3c0c8bb3898d617ddcda724403052b9a22
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/121553
Reviewed-by: dan sinclair <dsinclair@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>