Fix `kLinuxJpFontTable` row vs. column size confusion
Currently, `kLinuxJpFontTable` is a 4x4 array where there are 4 rows for
the 4 common Japanese font styles. In each row, the 4 entries are a list
of fonts of that style. The row selection enum has a `kCount` value that
is being used to represent the number of entries is a row. It is wrong
but just happens to work because the array is a square.

To fix this confusion, remove `kCount` and separately add a
`kJpFontFamilyRowEntryCount` constants. Then rename the enum from
`JpFontFamily` to `JpFontFamilyRowIndex` to make it even clearer. This
will make it possible to add a 5th column of font names in the near
future.

Along the way, remove "Linux" from constant names inside
fx_linux_impl.cpp, as that is implied.

Change-Id: Ie4d6cd5ee74eeeb74185e9fc02dd2d2d4958d1f0
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/121830
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tom Sepez <tsepez@google.com>