Introduce FixedSizeDataVector<>::TryUninit()
Add a creation method for the uninitialized case that mirrors the "try"
method for the zeroed case. This is expected to be used in a follow-on
CL to replace some raw ptrs and FX_Alloc()s.

In turn, create pdfium::internal functions needed to support this.
Also add Alloc2D() for consistency with other functions.

Change-Id: I4719bc9d460a69aa1b186913777cb92e8bc2c743
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/122191
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tom Sepez <tsepez@google.com>