Introduce FixedSizeDataVector<>::TruncatedFrom().
Analogous to base::HeapArray<>::take_from().

This is very similar to the move constructor, except that it gives a
new vector that claims a smaller size.  When the memory is actually
freed, the allocator knows the correct size itself from the allocation
itself despite the smaller claimed size.

This allows the pattern where we guess the size of the vector, write
some data to it, and then reset the end to match what was written.

In keeping with the idea that sizes should be as immutable as
possible, require an rvalue-reference as an argument, so that the
old vector becomes moved-from after creating the new one.

Change-Id: I88656e184ac022af1b4fcfce651f6ff3c7bafb28
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/122192
Reviewed-by: Tom Sepez <tsepez@google.com>
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>