Fix FXDIB_Format::kRgb32 handling when drawing with Skia
For bitmaps of format `kRgb32`, PDFium expects the alpha channel to be
ignored. As such, sometimes the alpha value is 0xFF, but not always. On
the other hand, Skia's equivalent would be "kBGR_888x_SkColorType", but
that color type does not exist. Skia only has `kRGB_888x_SkColorType`.
The Skia alternative is to use `kBGRA_8888_SkColorType` with
`kOpaque_SkAlphaType`, but this setup requires the alpha channel to
always be 0xFF. To meet the requirement, change
CFX_DIBBase::RealizeSkImage() to handle `FXDIB_Format::kRgb32` with a
transform that always sets the alpha channel to 0xFF. Then the rest of
the PDFium code can stay the same, without having to find every `kRgb32`
usage and set the alpha channel.

With this fix, the known to be bad test expectation for
transfer_function.in on Mac/Skia can be removed, and rendering this file
with Skia is consistent across platforms.

Bug: 351796783
Change-Id: I9308573c545a55637874450df98b2a79dad0bbe3
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/122350
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tom Sepez <tsepez@google.com>