Make FFLCommon() more robust with absl::variant
FFLCommon() is a helper function that is intended to either takes a
bitmap or a canvas. Currently, the bitmap and canvas are passed in as
two separate parameters, and they may be null. As a result, FFLCommon()
is not quite sure which to use, and tries to use both when initializing
CFX_DefaultRenderDevice.

Combine the 2 parameters into an absl::variant. Now FFLCommon() knows
for certain what the input type is, and can initialize
CFX_DefaultRenderDevice only once. Also tweak the contract with
FFLCommon(), so the callers always pass in a non-null bitmap/canvas.

Change-Id: I12c70a933f8eeb60c7afae1973bcc9614681730c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/123011
Reviewed-by: Tom Sepez <tsepez@google.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>