Make CFX_DIBBase::IsPremultiplied() available in non-Skia builds
Make it available so //fpdfsdk code can more easily check that
FPDF_BITMAPs that are passed in and being returned are not
pre-multiplied. This is important to ensure right now, as PDFium does
not have a public bitmap type that corresponds to
`FXDIB_Format::kArgbPremul` yet, and upcoming CLs will try to put that
format to use.

Bug: 42271020
Change-Id: Ife31dea4d31d3fbd64f743e8a82b95ef611839de
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/122933
Reviewed-by: Tom Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>