Add CFX_DIBitmap::ScopedPremultiplier
Add a scoper that calls CFX_DIBitmap::PreMultiply() and UnPreMultiply().
This will be used to help ensure bitmaps passed into PDFium that get
pre-multiplied are un-premultiplied when returned to the caller.

Since CFX_DIBitmap::PreMultiply() did not exist yet, add it as a stub.

Bug: 42271020
Change-Id: I45ca2b83889d9520c6e504dd395fa957ed57be5b
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/122913
Reviewed-by: Tom Sepez <tsepez@google.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>