Make CFX_DIBitmap::GetScanlineAs() work correctly for 24-bpp bitmaps
Add a test case that demonstrates CFX_DIBitmap allocates memory with a
stride of 12 when the bitmap width is 3 and the format is 24-bpp. The
test case then shows the 12 bytes gets reinterpreted as 4 pixels of 3
bytes each by GetScanlineAs(). Fix GetScanlineAs() to only return the
actual pixels and not the unused space.

Do the same for GetWritableScanlineAs().

Change-Id: I3181bc5dbb5175b91671a995168ff3be21615389
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/123210
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@google.com>