Adjust font fallback code to prefer embedded fonts
When a PDF already embeds a font, ShouldUseExistingFont() should return
true. Thus preferring that over the fallback font. This fixes a text
rendering issue where the font is embedded, but is not being used. Add a
pixel test for this case.

Also update the rendering for bug_1388_3.in, which now displays a tofu
box. Several other PDF implementations also render this tofu box.

Bug: 358039783
Change-Id: Icf6244dc480183fee54055d49f58660d745ac089
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/123390
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@google.com>
Reviewed-by: Tom Sepez <tsepez@chromium.org>