Validate sRGB ICC profiles have 3 components
CPDF_IccProfile currently marks ICC profiles as sRGB profiles based on
whether the data has the sRGB marker. Improve the validation by also
checking whether the profiles have 3 components or not.

Add a pixel test for this case based on the existing
icc_profile_bad_value.in test. Without this CL, the test file renders as
all red, and with this CL, it renders as all white.

Bug: 361712478
Change-Id: I3a97d7ba33033105834b01ce06a750b90a692e19
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/124270
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Tom Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>