Set a memory limit for libtiff
Fuzzers can easily generate output that causes libtiff to allocate a lot
of memory, which then causes the fuzzer to fail. The fuzzer's bitmap
pixel limit is set to 64 MB, and lowering that does not help. So instead
of lowering the limit there, restrict how much memory libtiff can
allocate to make the fuzzers happy. The limit can be adjusted later if
it turns out to be an issue in the real world.

Bug: 349530605
Change-Id: I06e9bd31adc08d556cefb397f8b11eef1eea074c
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/125390
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>