Remove METADATA file from libtiff.
This previously had a list of mitigated CVE numbers in it, because they
did not apply to our usage of libtiff. These have now been removed, so
we can delete this file. If future CVEs need to be excluded from
scanning they can be added to the README file as this is now supported
and preferred.

Bug: 373044593
Change-Id: I40c897982b34ea80c7fffae028b0255c8dd961fc
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/125190
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Auto-Submit: Rachael Newitt <renewitt@google.com>