Make PDFium's compiler_specific.h use clang's UNSAFE_BUFFERS_BUILD
This symbol is defined consistently in build/config/clang/BUILD.gn
and is a better source of truth than the old PDFium symbol.

This allows the UNSAFE_BUFFERS macro to be correctly defined even
outside of pdfium_common_config, such as for fuzzers and testing.

Change-Id: I5f15147e8eeb28104323b5a58f09707a8df6de7e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/127730
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>
Commit-Queue: Tom Sepez <tsepez@chromium.org>