[AGG] Only add positive dash lengths and gap lengths
It does not make sense for a dash length or gap length to be negative,
and a CHECK is failing because of negative dash lengths being added in
agg::vcgen_dash::add_dash(). Modify the only caller to make sure the
values being passed in are always positive.

Bug: 370299908
Change-Id: I027411a6a64e07fd392ca6f8211617cfca0b631f
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/127350
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>