Switch FXSYS_wcstof() to use fast_float
The existing FXSYS_wcstof() is rather inefficient, as it has to make a
copy of the input to ensure it is NUL-terminated. Switch to using
fast_float, which may use a faster algorithm, but is also faster because
it does not require string copying. This makes a pdf_formcalc_fuzzer
input that ends up calling FXSYS_wcstof() 100X faster.

Bug: 392928937
Change-Id: I2092522487d9373e454f870bfefd409fe96f4f99
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/128471
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>