Refine NeedToPremultiplyBitmap() logic
CFX_DIBitmap::ScopedPremultiplier::NeedToPremultiplyBitmap() determines
if a bitmap needs to be pre-multiplied, yet it does not check the bitmap
at all. This is because it is relying on two crutches:

1) CFX_DIBitmap::PreMultiply() and CFX_DIBitmap::UnPreMultiply() are
   no-ops for formats without alpha.
2) It never handles pre-multiplied bitmaps.

Condition (2) will go away in the future, so anticipate that and change
NeedToPremultiplyBitmap() to check the bitmap's format.

Bug: 42271033
Change-Id: I4d961c5fa20bd254962562e6ce51ef4236ef7360
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/129012
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Thomas Sepez <tsepez@google.com>