Change FPDFImageObj_GetRenderedBitmap() to take clip path into account
If there exists a clip path for a bitmap, then only part of the bitmap
should be rendered. Change FPDFImageObj_GetRenderedBitmap() to take the
clip path into account. To do so, the clip path has to be transformed to
match the image:

1) The image may have been transformed by PDFium APIs.
2) The image has to be transformed from where it is on the page to fit
   into the output bitmap.

After which, the clip path needs to be transformed one more time to be
flipped around, which is what CPDF_RenderStatus::ProcessClipPath()
expects, similar to CPDF_ImageRenderer::Start().

Bug: 406540676
Change-Id: I28eca20d4becf799be54d0cf093af31ba7f4c4f7
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/130450
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>