Clamp PDFium indexed color space hival
The PDF spec indicates that the high value for an indexed color space
must be in the range of 0 to 255.  Should a color space violate this,
failing out of loading the color space results in images which rely
on the color space to not load, and thus not be displayed by viewers.

Some PDF viewers (but not all) do allow for this out-of-range value,
and take a corrective interpretation so that such images can be loaded.

Modify the approach to instead clamp the specified high value to be
within the spec range limit.  This slight corrective action allows such
images to now be loaded by PDFium.

Fixed: 372749732
Change-Id: Ib77d46d04e9da6dbc0e9345653c38241a15dc45e
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/129951
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>