Handle arrays of 1-in, 1-out functions for shading types 4-7
Shading types 1-3 have similar code in cpdf_rendershading.cpp already.

Since there's no tests at all for several shading types, add
a test that uses each shading type, both with a 1-in 3-out function
and an array of 1-in 1-out functions.

Due to crbug.com/406662638, the expected file for the last 3 shading
types is missing a visible color stop (i.e. this is also a test for
that bug). But before this change, the last 4 shading types drew just
red for the last 4 steps for the array case. Now the array case has the
same output as the 1-on 3-out case.

Bug: 413635900
Change-Id: I6c46f1b8164cfee8e07126deac5da4f99de4d5f5
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/131090
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Nico Weber <thakis@google.com>
Commit-Queue: Lei Zhang <thestig@chromium.org>