Current font need not contain part of ActualText
When handling marked content there is currently a check that if the
current font does not have a mapping for any of the code points in an
ActualText then the ActualText is ignored. This check has been in the
code since the initial code drop. This undermines part of the purpose of
ActualText, such as providing the original code points even if there was
no font which represented them.

In addition, it is not clear why this behavior would be wanted. One
possibility is an attempt to reduce phishing where the copied (or
accessible) content is intentionally different from was is drawn on the
page (especially with respect to Unicode confusables). However, this is
generally malicious content which could easily work around such a check
by simply adding mappings in its fonts. Any benefit of ignoring such
ActualText has the downside of incorrectly handling real and intentional
uses like when multiple code points map to a single glyph in a font.

Remove this check and add a test which replaces English words with
Chinese replacements. This test fails before this change as the default
test font does not contain both sets of code points.

Bug: 420508260
Change-Id: I4d82b603e5bda8ed13614e6ba3a4ce0c8f8f0d62
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/132710
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>