Fix undo queue state for ReplaceSelection items
The DCHECK crash bug was introduced as a part of a bug fix in CL [1].

CL [1] introduced the scheme where Undo/Redo returns the number of items
remaining and everything but UndoReplaceSelection returned 0. The
original problem that needed fixing was that ReplaceSelection inserts
multiple items into the undo queue and all must be executed with undo
and redo operations.

Initially, the number of undo items for ReplaceSelection was hard-coded
to 3. However this was buggy because the number of items is actually
variable. So CL [2] added an additional complication in ReplaceSelection
where the internal logic of the other helpers was replicated to
determine the number of items inserted into the queue. This number was
intended to be set on both UndoReplaceSelection objects via
UndoStack::GetLastAddItem().

The fix in CL [2] had two problems:

 1. The replication of the internal logic was imperfect, so the count
    was sometimes wrong. It was even revised later but it was still
    wrong.
 2. GetLastAddItem() isn't a reliable way to access the
    UndoReplaceSelection object. Sometimes set_undo_remaining() was
    called on other item types that ignore it.

This CL takes a different approach: don't try to set these item counts
at all. Instead simply look for the two UndoReplaceSelection items and
execute every item in between.

[1] https://pdfium.googlesource.com/pdfium/+/3abc96a18bf41a55958e2fe99a3ce03882c3ecb5
[2] https://pdfium.googlesource.com/pdfium/+/6a34da391b15f5373247fda48c10a12bc4143c94

Bug: 445725426
Change-Id: I61fc85d022200574f93c981150471b27150ca492
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/135850
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: April Kallmeyer <ask@chromium.org>