Fix undo queue limit cut-off with multi-item undos
Since ReplaceSelection adds a variable number of items to the undo
queue, and the queue has a limit, it is possible to end up in a
situation where only part of the ReplaceSelection actions are left in
the queue.

So make sure to remove all of the related undo items when trimming the
queue.

Bug: 445725426
Change-Id: I256348571a9c37692cc6c7084979161339022927
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/135851
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: April Kallmeyer <ask@chromium.org>