Add type tags to calls to v8::External
The v8::External API got extended to take a type tag parameter. A
v8::External created with a Tag A can only be unwrapped with the same
Tag A.

So far, this CL only adds the default tag everywhere in PDFium.
Ideally, each type stored in a v8::External would get its own tag.
However, it is unclear so far how to synchronize the tags used in
PDFium with the tags used in Blink.

Change-Id: Ic18c72f6c008c2f31796b9a48b9a3625a3ca7ca2
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/136650
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Tom Sepez <tsepez@chromium.org>