Remove redundant jbig2 test files
jbig2_mmr.in is the same file as bitmap-mmr.jbig in
https://pdfium-review.googlesource.com/c/pdfium_tests/+/137990
(except that the .in file uses a /ASCIIHexDecode filter, which is
irrelevant to this test).

jbig2-trailing-7fff-stripped-harder{,-refine}.in is the same as
bitmap-trailing-7fff-stripped-harder{,-refine}.pdf over there.

No behavior change.

Bug: 42271957
Change-Id: I3f2ff04309f5effa06eedb1aaa6775cb2deabb47
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/138070
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Nico Weber <thakis@chromium.org>
Reviewed-by: Nico Weber <thakis@google.com>