Update CFX_Face::Open() Callers in CFPF_SkiaFontMgr
This CL creates a new function CFX_Face::OpenFromFilePath() which
constructs FT_Open_Args which is called in CFPF_SkiaFontMgr. This is
done to avoid arg construction in the Callee and is a continuation of
the effort to consolidate CFX_Face::Open() Callers and eventually make
the previously mentioned method private.

Bug: 459848480
Change-Id: Ieeeccb66342a34b6c5826927ce6016449e93dcca
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/138250
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>