Overwrite the timezone for JS/pixel/corpus tests on Linux
Set the TZ environment variable in test_runner.py, so all test suites
that use it run tests in the same TZ. Use America/Los_Angeles, AKA
PST/PDT, as that is what the existing test expectations assume.

For now, just do this on Linux, since that is the platform currently
failing the continuous build. Whereas trying to do this on Windows seems
to have the opposite effect and actually triggers the issue.

Bug: 461544934
Change-Id: I7ed7e02fe14b1440cb45e15e66d9ba4d3212c0fb
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/138410
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>